requirements.txt
Flask==3.0.2
playwright==1.42.0
weasyprint==60.2
gitpython==3.1.42
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.3 